enable -Wconversion and fix the occurring warnings

+++ periph/cortexm4f/scb/scb.cpp
@@ -28,7 +28,7 @@ CpuInfo SystemControlBlock::get_cpu_info() const noexcept
     CpuInfo info{};
     uint32_t raw = reg().cpuid.get();
 
-    info.impl_code = (raw >> 24) & 0xFF;
+    info.impl_code = static_cast<uint8_t>((raw >> 24) & 0xFF);
     info.variant = (raw >> 20) & 0x0F;
     info.constant = (raw >> 16) & 0x0F;
     info.part_number = (raw >> 4) & 0xFFF;
